% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{value_by_year}
\alias{value_by_year}
\title{Filter bees total value production on multiple years}
\usage{
value_by_year(dataset, state, year)
}
\arguments{
\item{dataset}{A data frame containing the bee colony information.}

\item{state}{A character string specifying the state for which to filter the data.}

\item{year}{A numeric or vector of years to filter the data (e.g., 1995 or c(1995, 1996)).}
}
\value{
A data frame containing the total value of production for the specified `state` and `year`.
}
\description{
This function allows the user to get the total value of production for one or more years for a specific state in the USA.
}
\details{
This function allows the user to get the total value production for one or multiple years for a specific state in the USA.
}
\examples{

# Get value total for Alabama state on 1995
value_by_year(honey_data, "Alabama", 1995)

# Get value total for Alabama state on 1995 and 1996
value_by_year(honey_data, "Alabama", c(1995, 1996))
}
